[refactor] MySQL 전환 이전 코드들 정리(Study, User)
